+++ src/pyobo/__init__.py
@@ -55,7 +55,7 @@ from .api import (
     is_descendent,
 )
 from .getters import get_ontology
-from .normalizer import OboNormalizer, ground
+from .ner.normalizer import OboNormalizer, ground
 from .obographs import parse_results_from_obo
 from .plugins import (
     has_nomenclature_plugin,
